Updated the AMI service behavior to disclose welcome messages and other settings to clients. Updated the app service to save configurations updated the web error behavior to emit the content type. Updated the configuration providers to emit their configuration to the context when the UpdateConfiguration method is called

+++ SanteDB.Rest.AppService/Configuration/RealmConfigurationFeature.cs
@@ -89,7 +89,7 @@ namespace SanteDB.Rest.AppService.Configuration
             section.Realm = new UpstreamRealmConfiguration()
             {
                 DomainName = featureConfiguration[REALM_NAME]?.ToString(),
-                PortNumber = (int?)featureConfiguration[PORT_NUMBER] ?? 8080,
+                PortNumber = (int)((long?)featureConfiguration[PORT_NUMBER] ?? 8080),
                 UseTls = (bool?)featureConfiguration[USE_TLS] ?? false
             };
